# doc/openapi/brokerage-api.openapi.yml
openapi: 3.0.3
info:
  title: Insurance Brokerage API
  version: 1.1.0
  description: OpenAPI contract for brokerage P0 and P1 endpoints (party KYC, broker license, distribution agreement, product management, visibility, broker offerings, catalog BFF).
servers:
  - url: http://localhost:18000/api/v1
paths:
...
  /products:
    post:
      summary: Create a product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '200':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    get:
      summary: List products with filtering and pagination
      operationId: listProducts
      parameters:
        - name: lineOfBusiness
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [draft, active, retired, archived] }
        - name: ownerOrganizationId
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /products/{productId}:
    get:
      summary: Get a product by ID
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /products/{productId}/versions:
    post:
      summary: Create a new product version
      operationId: createProductVersion
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductVersionCreateRequest'
      responses:
        '200':
          description: Version created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    get:
      summary: List product versions
      operationId: listProductVersions
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: List of versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /products/{productId}/versions/{version}:
    get:
      summary: Get a specific product version
      operationId: getProductVersion
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: version
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Product version details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /products/{productId}/versions/{version}/activate:
    post:
      summary: Activate a product version (supersedes previous active version)
      operationId: activateProductVersion
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: version
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object, properties: { effectiveFrom: { type: string, format: date-time } } }
      responses:
        '200':
          description: Version activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /products/{productId}/versions/{version}/retire:
    post:
      summary: Retire a product version
      operationId: retireProductVersion
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: version
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Version retired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /products/{productId}/versions/{version}/clone:
    post:
      summary: Clone a product version to create a new draft
      operationId: cloneProductVersion
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: version
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Version cloned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  # ---- P1-2 Product Visibility ----
  /products/{productId}/visibility:
    post:
      summary: Create product visibility for a distributor
      operationId: createProductVisibility
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductVisibilityCreateRequest'
      responses:
        '200':
          description: Visibility created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    get:
      summary: List product visibilities
      operationId: listProductVisibilities
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: List of visibilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /products/{productId}/visibility/{visibilityId}:
    get:
      summary: Get a product visibility
      operationId: getProductVisibility
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: visibilityId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Visibility details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    patch:
      summary: Update a product visibility
      operationId: updateProductVisibility
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: visibilityId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Visibility updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /products/{productId}/visibility/{visibilityId}/revoke:
    post:
      summary: Revoke a product visibility
      operationId: revokeProductVisibility
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: visibilityId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Visibility revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /distributors/{distributorOrganizationId}/visible-products:
    get:
      summary: List visible products for a distributor
      operationId: listDistributorVisibleProducts
      parameters:
        - name: distributorOrganizationId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: productVersion
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: List of visible products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  # ---- P1-3 Broker Product Offering ----
  /broker-offerings:
    post:
      summary: Create a broker product offering
      operationId: createBrokerProductOffering
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrokerProductOfferingCreateRequest'
      responses:
        '200':
          description: Offering created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    get:
      summary: List broker product offerings
      operationId: listBrokerProductOfferings
      parameters:
        - name: brokerOrganizationId
          in: query
          schema: { type: string, format: uuid }
        - name: lineOfBusiness
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: List of offerings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /broker-offerings/{offeringId}:
    get:
      summary: Get a broker product offering
      operationId: getBrokerProductOffering
      parameters:
        - name: offeringId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Offering details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    patch:
      summary: Update a broker product offering
      operationId: updateBrokerProductOffering
      parameters:
        - name: offeringId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '200':
          description: Offering updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /broker-offerings/{offeringId}/activate:
    post:
      summary: Activate a broker product offering
      operationId: activateBrokerProductOffering
      parameters:
        - name: offeringId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Offering activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /broker-offerings/{offeringId}/inactivate:
    post:
      summary: Inactivate a broker product offering
      operationId: inactivateBrokerProductOffering
      parameters:
        - name: offeringId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Offering inactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /customers/offerings:
    get:
      summary: List customer-facing offerings with ranking and disclosure
      operationId: listCustomerOfferings
      parameters:
        - name: brokerOrganizationId
          in: query
          schema: { type: string, format: uuid }
        - name: lineOfBusiness
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Customer offerings with commission disclosure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  # ---- P1-5 Catalog BFF ----
  /catalog/products:
    get:
      summary: List products via catalog BFF (view model only, no rate table)
      operationId: catalogListProducts
      parameters:
        - name: lineOfBusiness
          in: query
          schema: { type: string }
        - name: ownerOrganizationId
          in: query
          schema: { type: string, format: uuid }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Catalog product list (view model)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /catalog/products/{productId}:
    get:
      summary: Get product details via catalog BFF
      operationId: catalogGetProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Product view model
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  schemas:
...
    ProductCreateRequest:
      type: object
      required: [productCode, nameFa, lineOfBusiness]
      properties:
        productCode: { type: string }
        nameFa: { type: string }
        nameEn: { type: string, nullable: true }
        lineOfBusiness: { type: string }
        effectiveFrom: { type: string, format: date-time, nullable: true }
        effectiveTo: { type: string, format: date-time, nullable: true }
    ProductVersionCreateRequest:
      type: object
      required: [changeReason]
      properties:
        changeReason: { type: string }
        nameFa: { type: string, nullable: true }
        nameEn: { type: string, nullable: true }
        effectiveFrom: { type: string, format: date-time, nullable: true }
        formSchema: { type: object, nullable: true }
        requiredDocuments: { type: array, items: { type: string }, nullable: true }
        coverages:
          type: array
          items: { type: object }
        rateTables:
          type: array
          items: { type: object }
    ProductVisibilityCreateRequest:
      type: object
      required: [distributionAgreementId, agreementVersionAtCreation, visibilityType, effectiveFrom]
      properties:
        distributorOrganizationId: { type: string, format: uuid, nullable: true }
        distributionAgreementId: { type: string, format: uuid }
        agreementVersionAtCreation: { type: integer }
        visibilityType: { type: string, enum: [private, exclusive, marketplace] }
        markupRules: { type: array, items: { type: object }, nullable: true }
        allowedTerritories: { type: array, items: { type: string } }
        allowedSalesChannels: { type: array, items: { type: string } }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: string, format: date-time, nullable: true }
    BrokerProductOfferingCreateRequest:
      type: object
      required: [name, includedProductIds, distributionAgreementId, agreementVersionSnapshot, effectiveFrom]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        includedProductIds: { type: array, items: { type: string, format: uuid } }
        markupRules: { type: array, items: { type: object }, nullable: true }
        allowedSalesChannels: { type: array, items: { type: string } }
        distributionAgreementId: { type: string, format: uuid }
        agreementVersionSnapshot: { type: integer }
        effectiveFrom: { type: string, format: date-time }
        effectiveTo: { type: string, format: date-time, nullable: true }
...
    SuccessResponse:
      type: object
      required: [success, data, correlationId]
      properties:
        success: { type: boolean }
        data: {}
        correlationId: { type: string }
...